11/26 1:01am
Not much changed since the last time saved- just a couple added debug statements. Seems like gameBoard is working just fine, but the aliasing on the white pieces especially is wonky as hell. Not sure what's going on, but I would assume something to do with the eval statements (as I said prior). going to have to work through the code pretty meticulously and do like code tracing to get that fixed. Going to bed fr this time.